change to use int + add comment

+++ NorthstarDLL/modmanager.cpp
@@ -470,12 +470,14 @@ void ModManager::LoadMods()
 					modPak.m_sPakName = pakName;
 
 					// read header of file and get the starpak paths
+					// this is done here as opposed to on starpak load because multiple rpaks can load a starpak
+					// and there is seemingly no good way to tell which rpak is causing the load of a starpak :/
 
 					std::ifstream rpakStream(file.path(), std::ios::binary);
 
 					// seek to the point in the header where the starpak reference size is
 					rpakStream.seekg(0x38, std::ios::beg);
-					uint16_t starpaksSize;
+					int starpaksSize;
 					rpakStream.read((char*)&starpaksSize, 2);
 
 					// seek to just after the header
@@ -488,7 +490,7 @@ void ModManager::LoadMods()
 
 					// split the starpak reference(s) into strings to hash
 					std::string str = "";
-					for (uint16_t i = 0; i < starpaksSize; i++)
+					for (int i = 0; i < starpaksSize; i++)
 					{
 						if (buf[i] != 0x00)
 						{
